[FLINK-31721][core] Move JobStatusHook to flink-core module
